Move generic API traits and classes to Api/Generic namespace
