+++ src/BigCommerce/Api/Generic/DeleteInIdList.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BigCommerce\ApiV3\Api;
+namespace BigCommerce\ApiV3\Api\Generic;
 
 use BigCommerce\ApiV3\Client;
 use GuzzleHttp\RequestOptions;
