+++ src/BigCommerce/Api/Generic/DeleteResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BigCommerce\ApiV3\Api;
+namespace BigCommerce\ApiV3\Api\Generic;
 
 use BigCommerce\ApiV3\Client;
 use Psr\Http\Message\ResponseInterface;
