+++ src/BigCommerce/Catalog/Brands/BrandImageApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog\Brands;
 
-use BigCommerce\ApiV3\Api\ResourceImageApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceImageApi;
 
 class BrandImageApi extends ResourceImageApi
 {
