+++ src/BigCommerce/Catalog/Brands/BrandMetafieldsApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog\Brands;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\ResponseModels\Brand\BrandMetafieldResponse;
 use BigCommerce\ApiV3\ResponseModels\Brand\BrandMetafieldsResponse;
