+++ src/BigCommerce/Catalog/BrandsApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\Catalog\Brands\BrandImageApi;
 use BigCommerce\ApiV3\Catalog\Brands\BrandMetafieldsApi;
 use BigCommerce\ApiV3\ResponseModels\Brand\BrandResponse;
