+++ src/BigCommerce/Catalog/CategoriesApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\Catalog\Categories\CategoryImageApi;
 use BigCommerce\ApiV3\Catalog\Categories\CategoryMetafieldsApi;
 use BigCommerce\ApiV3\ResourceModels\Catalog\Category\Category;
