+++ src/BigCommerce/Catalog/Products/ComplexRulesApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog\Products;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\ResourceModels\Catalog\Product\ComplexRule;
 use BigCommerce\ApiV3\ResponseModels\ComplexRuleResponse;
 use BigCommerce\ApiV3\ResponseModels\ComplexRulesResponse;
