+++ src/BigCommerce/Catalog/Products/ProductVideosApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Catalog\Products;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\ResponseModels\Product\ProductVideoResponse;
 use BigCommerce\ApiV3\ResponseModels\Product\ProductVideosResponse;
