+++ src/BigCommerce/Orders/OrderMetafieldsApi.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\Orders;
 
-use BigCommerce\ApiV3\Api\ResourceApi;
+use BigCommerce\ApiV3\Api\Generic\ResourceApi;
 use BigCommerce\ApiV3\ResponseModels\Order\OrderMetafieldResponse;
 use BigCommerce\ApiV3\ResponseModels\Order\OrderMetafieldsResponse;
