+++ src/BigCommerce/ResponseModels/Brand/BrandsResponse.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\ResponseModels\Brand;
 
-use BigCommerce\ApiV3\Api\FetchAllPages;
+use BigCommerce\ApiV3\Api\Generic\FetchAllPages;
 use BigCommerce\ApiV3\ResourceModels\Catalog\Brand\Brand;
 use BigCommerce\ApiV3\ResponseModels\PaginatedResponse;
