+++ src/BigCommerce/ResponseModels/Order/TransactionsResponse.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3\ResponseModels\Order;
 
-use BigCommerce\ApiV3\Api\FetchAllPages;
+use BigCommerce\ApiV3\Api\Generic\FetchAllPages;
 use BigCommerce\ApiV3\ResourceModels\Order\Transaction;
 use BigCommerce\ApiV3\ResponseModels\PaginatedResponse;
